Method to fix utf8 from any array
Thank you very much to your class, helped to solve the many problems in the project in which I am part. I've added a method to be used to fix arrays, which is very useful in json_encode's that return "Malformed UTF-8 characters, possibly incorrectly encoded". I hope to add a lot to this class that helped me a lot.